+++ CHANGELOG.rst
@@ -9,6 +9,7 @@ CHANGELOG
 * bug-fix: Change module names to string type in __all__
 * bug-fix: tensorflow-serving-api: SageMaker does not conflict with tensorflow-serving-api module version
 * feature: Local Mode: add support for local training data using file://
+* feature: Updated TensorFlow Serving api protobuf files
 
 1.2.4
 =====